Fix: Fixed renaming of namespaces in config files
