+++ PV260_Project/BusinessLayer/BusinessLayer.csproj
@@ -4,7 +4,6 @@
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>BL</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
