+++ PV260_Project/BusinessLayer/Exceptions/DataWriterException.cs
@@ -1,4 +1,4 @@
-namespace BL.Exceptions;
+namespace BusinessLayer.Exceptions;
 
 [Serializable]
 public class DataWriterException : Exception
