+++ PV260_Project/BusinessLayer/Logging/ILogger.cs
@@ -1,4 +1,4 @@
-namespace BL.Logging;
+namespace BusinessLayer.Logging;
 
 public interface ILogger
 {
