+++ PV260_Project/DataLayer/DataLayer.csproj
@@ -4,7 +4,6 @@
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>DAL</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
