+++ PV260_Project/PresentationLayer/IApp.cs
@@ -1,4 +1,4 @@
-namespace PL;
+namespace PresentationLayer;
 
 public interface IApp
 {
