+++ PV260_Project/PresentationLayer/PresentationLayer.csproj
@@ -5,7 +5,6 @@
         <TargetFramework>net6.0</TargetFramework>
         <ImplicitUsings>enable</ImplicitUsings>
         <Nullable>enable</Nullable>
-        <RootNamespace>PL</RootNamespace>
     </PropertyGroup>
 
     <ItemGroup>
