+++ PV260_Project/PresentationLayer/Program.cs
@@ -1,6 +1,6 @@
-using PL.ConsoleApps;
+using PresentationLayer.ConsoleApps;
 
-namespace PL
+namespace PresentationLayer
 {
     public class Program
     {
